src/requirements.txt
flask==1.1.1
pymysql==0.9.3
flask-sqlalchemy==2.4.1
requests==2.22.0
attrs==19.2.0
cattrs==1.0.0
loguru==0.4.1
beautifulsoup4==4.8.0
lxml==4.4.1
flask-cors==3.0.8
pika==1.1.0
y
